Move recently created classes to a more meaningful file
Events related classes moved from "brownie/network/alert.py" to "brownie/network/event.py"
