+++ brownie/network/alert.py
@@ -1,16 +1,10 @@
 #!/usr/bin/python3
 
-import queue
-import time as time
-from threading import Lock, Thread
+import time
+from threading import Thread
 from typing import Callable, Dict, List, Tuple, Union
 
-from web3._utils import filters
-from web3.datastructures import AttributeDict
-
 from brownie.utils import color
-
-from .web3 import ContractEvent, web3
 
 __console_dir__ = ["Alert", "new", "show", "stop_all"]
 _instances = set()
@@ -122,156 +116,6 @@ class Alert:
             self.wait()
 
 
-class EventWatchData:
-    def __init__(
-        self,
-        event: ContractEvent,
-        callback: Callable[[AttributeDict], None],
-        delay: float = 2.0,
-        repeat: bool = True,
-        from_block: int = None,
-    ) -> None:
-        # Args
-        self.event = event
-        self.callback = callback
-        self.delay = delay
-        self.repeat = repeat
-        # Members
-        self._event_filter: filters.LogFilter = event.createFilter(
-            fromBlock=(from_block if from_block is not None else web3.eth.block_number - 1)
-        )
-        self._cooldown_time_over: bool = False
-        self._update_trigger_time()
-
-    def check_timer(self) -> float:
-        current_time: float = time.time()
-        time_since_last_trigger: float = current_time - self.last_trigger_time
-        # Sets callback execution as ready when timer is greater than delay
-        if time_since_last_trigger >= self.delay:
-            self.cooldown_time_over = True
-        # Return time left on watch in seconds
-        return self.delay - time_since_last_trigger
-
-    def get_new_events(self) -> List["filters.LogReceipt"]:
-        return self._event_filter.get_new_entries()
-
-    def _trigger_callback(self, events_data: List[AttributeDict]) -> None:
-        self.cooldown_time_over = False
-        self._update_trigger_time()
-        for data in events_data:
-            self.callback(data)
-
-    def _update_trigger_time(self) -> None:
-        self.last_trigger_time: float = time.time()
-
-    def __cd_time_over_getter(self) -> bool:
-        current_time = time.time()
-        timer_value = current_time - self.last_trigger_time
-        if timer_value >= self.delay:
-            self._cooldown_time_over = True
-        return self._cooldown_time_over
-
-    def __cd_time_over_setter(self, value: bool) -> None:
-        self._cooldown_time_over = value
-
-    cooldown_time_over = property(fget=__cd_time_over_getter, fset=__cd_time_over_setter)
-
-
-class EventWatcher:
-    def __init__(self) -> None:
-        self.target_list_lock: Lock = Lock()
-        self.target_events_watch_data: List[EventWatchData] = []
-        self._queue: queue.Queue = queue.Queue()
-        self._kill: bool = False
-        self._has_started: bool = False
-        self._watcher_thread = Thread(target=self._watch_loop, daemon=True)
-        self._callback_thread = Thread(target=self._execute_callbacks, daemon=True)
-
-    def stop(self, wait: bool = True) -> None:
-        self._kill = True
-        if wait is True:
-            self._watcher_thread.join()
-            self._callback_thread.join()
-
-    def add_event_callback(
-        self,
-        event: ContractEvent,
-        callback: Callable[[AttributeDict], None],
-        delay: float = 2.0,
-        repeat: bool = True,
-        from_block: int = None,
-    ) -> None:
-        if self._has_started is False:
-            self._start_threads()
-        self._add_event_callback(event, callback, delay, repeat, from_block)
-
-    def _start_threads(self) -> None:
-        # Starts two new Thread running the _watch_loop and the _execute_callbacks method.
-        self._watcher_thread.start()
-        self._callback_thread.start()
-        self._has_started = True
-
-    def _add_event_callback(
-        self,
-        event: ContractEvent,
-        callback: Callable[[AttributeDict], None],
-        delay: float = 2.0,
-        repeat: bool = True,
-        from_block: int = None,
-    ) -> None:
-        if not callable(callback):
-            raise TypeError("Argument 'callback' argument must be a callable.")
-        delay = max(delay, 0.05)
-        self.target_list_lock.acquire()  # lock
-        self.target_events_watch_data.append(
-            EventWatchData(event, callback, delay, repeat, from_block)
-        )
-        self.target_list_lock.release()  # unlock
-
-    def _execute_callbacks(self) -> None:
-        while not self._kill:
-            try:
-                print("[EXECUTER] - SafeQueue size : {}. Executing...".format(self._queue.qsize()))
-                while self._queue.qsize() > 0:
-                    # @dev: Not using Queue.get method for cross-platform reasons.
-                    #   @see: https://docs.python.org/3/library/queue.html#queue.Queue.get
-                    # Raises queue.Empty exception if queue is empty
-                    task_data = self._queue.get_nowait()
-                    # Execute callbacks with new events data
-                    task_data["function"](task_data["events_data"])
-            except queue.Empty:
-                pass
-            # Sleep a few before checking for new events
-            time.sleep(0.1)
-
-    def _watch_loop(self) -> None:
-        while not self._kill:
-            try:
-                # print("[WATCHER] - Awake ! Checking...")
-                sleep_time: float = 2.0  # Max sleep time.
-                self.target_list_lock.acquire()  # lock
-                for elem in self.target_events_watch_data:
-                    # print(f"[WATCHER] - Watching event {elem.event.event_name}")
-                    # If cooldown is not over, skip.
-                    if elem.cooldown_time_over is False:
-                        sleep_time = min(sleep_time, elem.check_timer())
-                        continue
-                    # print("[WATCHER] - Cooldown reached ! Checking for new events...")
-                    # Check for new events & execute callback async if some are found
-                    latest_events = elem.get_new_events()
-                    if len(latest_events) != 0:
-                        self._queue.put(
-                            {
-                                "function": elem._trigger_callback,
-                                "events_data": latest_events,
-                            }
-                        )
-                    sleep_time = min(sleep_time, elem.delay)
-            finally:
-                self.target_list_lock.release()  # unlock
-                time.sleep(sleep_time)
-
-
 def new(
     fn: Callable,
     args: Tuple = None,
@@ -295,6 +139,3 @@ def stop_all() -> None:
     for t in _instances.copy():
         t.stop()
     _instances.clear()
-
-
-event_watcher = EventWatcher()
